requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
filelock==3.6.0
Flask==1.1.2
gevent==21.8.0
greenlet==1.1.1
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.1.0
MarkupSafe==1.1.1
numpy==1.21.5
packaging==21.3
Pillow==9.1.0
pip==21.2.4
pycparser==2.21
pyparsing==3.0.7
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
setuptools==58.0.4
six==1.16.0
tokenizers==0.12.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.1
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==1.0.1
wheel==0.37.1
wincertstore==0.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
